Add dry run default to check some of the values

+++ lib/tasks/email_and_unsubscribe_archived_topics.rake
@@ -1,12 +1,16 @@
 namespace :archived_topics do
   desc "Send a bulk email to subscribers to these archived Specialist Topics."
-  task :send_emails, [] => :environment do |_t, _args|
+  task :email_and_unsubscribe, [:run] => :environment do |_t, args|
+    args.with_defaults(run: "true")
     topic_urls.each do |topic|
       sub_list = SubscriberList.find_by(url: topic[:url])
       if !sub_list
         puts "No SubscriberList for #{topic[:url]}"
       else
         puts "Sending email for #{sub_list.url} (ID: #{sub_list.id})"
+        puts "#{sub_list.subscribers.count} in this list"
+
+        subject = "Update from GOV.UK for: #{sub_list.title}"
 
         body = <<~BODY
           Update from GOV.UK for:
@@ -22,22 +26,33 @@ namespace :archived_topics do
 
           This topic has been archived. You will not get any more emails about it.
 
-          You can find more information about this topic at [#{topic[:redirect_title]}](#{topic[:redirect_url]}).
           You can find more information about this topic at [#{topic[:redirect_title]}](#{topic[:redirect]}).
         BODY
 
-        email_ids = BulkSubscriberListEmailBuilder.call(
-          subject: "Update from GOV.UK for: #{sub_list.title}",
-          body: body,
-          subscriber_lists: [sub_list],
-        )
+        puts "==============="
+        puts "DRY RUN" unless args[:run] != "true"
+        puts "CHECK OUTPUT"
+        puts "First SubscriberList..."
+        puts "The subscribers will see:"
+        puts "subject: #{subject}"
+        puts "body: #{body}"
+        puts "First recipient #{sub_list.subscribers.first.address}" if sub_list.subscribers.count.positive?
+        puts "==============="
 
-        email_ids.each do |id|
-          SendEmailWorker.perform_async_in_queue(id, queue: :send_email_immediate)
+        unless args[:run] == "true"
+          email_ids = BulkSubscriberListEmailBuilder.call(
+            subject: subject,
+            body: body,
+            subscriber_lists: [sub_list],
+          )
+
+          email_ids.each do |id|
+            SendEmailWorker.perform_async_in_queue(id, queue: :send_email_immediate)
+          end
+
+          puts "Destroying subscription list #{sub_list.url} (ID: #{sub_list.id})"
+          sub_list.destroy!
         end
-
-        puts "Destroying subscription list #{sub_list.url} (ID: #{sub_list.id})"
-        sub_list.destroy!
       end
     end
   end
